+++ fedn/fedn/clients/reducer/statestore/mongoreducerstatestore.py
@@ -70,9 +70,10 @@ class MongoReducerStateStore(ReducerStateStore):
                             # TODO Fix the ugly latering of indirection due to a bug in secure_filename returning an object with filename as attribute
                             # TODO fix with unboxing of value before storing and where consuming.
                             self.control.config.update({'key': 'package'},
-                                                        {'$set': {'filename': {'filename': control['context']}}}, True)
+                                                        {'$set': {'filename': control['context']}}, True)
                         if "helper" in control:
-                            self.set_framework(control['helper'])
+                            #self.set_framework(control['helper'])
+                            pass
 
                         round_config = {'timeout':180, 'validate':True}
                         try:
@@ -164,13 +165,12 @@ class MongoReducerStateStore(ReducerStateStore):
         except (KeyError, IndexError):
             return None
 
-    def set_framework(self,helper):
-        self.control.config.update({'key': 'framework'},
+    def set_framework(self, helper):
+        self.control.config.update({'key': 'package'},
                                     {'$set': {'helper': helper}}, True)
 
-
     def get_framework(self):
-        ret = self.control.config.find({'key': 'framework'})
+        ret = self.control.config.find({'key': 'package'})
         try:
             retcheck = ret[0]['helper']
             if retcheck == '' or retcheck == ' ':  # ugly check for empty string